Restructure test cases
