+++ tests/test_flows/test_simple_flow.py
@@ -1,0 +1,75 @@
+from __future__ import annotations
+
+import pandas as pd
+import sqlalchemy as sa
+
+from pydiverse.pipedag import Blob, Flow, Stage, Table, materialize
+
+
+@materialize(nout=2, version="1")
+def inputs():
+    dfA = pd.DataFrame(
+        {
+            "a": [0, 1, 2, 4],
+            "b": [9, 8, 7, 6],
+        }
+    )
+
+    dfB = pd.DataFrame(
+        {
+            "a": [2, 1, 0, 1],
+            "x": [1, 1, 2, 2],
+        }
+    )
+
+    import time
+
+    time.sleep(1)
+    return Table(dfA, "dfA"), Table(dfB, "dfA_%%")
+
+
+@materialize(input_type=pd.DataFrame)
+def double_values(df: pd.DataFrame):
+    return Table(df.transform(lambda x: x * 2))
+
+
+@materialize(input_type=sa.Table, lazy=True)
+def join_on_a(left: sa.Table, right: sa.Table):
+    return Table(left.select().join(right, left.c.a == right.c.a))
+
+
+@materialize(input_type=pd.DataFrame)
+def list_arg(x: list[pd.DataFrame]):
+    assert isinstance(x[0], pd.DataFrame)
+    return Blob(x)
+
+
+@materialize
+def blob_task(x, y):
+    return Blob(x), Blob(y)
+
+
+def test_simple_flow():
+    with Flow("FLOW") as flow:
+        with Stage("stage 1"):
+            a, b = inputs()
+            a2 = double_values(a)
+            b2 = double_values(b)
+            b4 = double_values(b2)
+            b4 = double_values(b4)
+            x = list_arg([a2, b, b4])
+
+        with Stage("stage 2"):
+            xj = join_on_a(a2, b4)
+            a = double_values(xj)
+
+            v = blob_task(x, x)
+            v = blob_task(v, v)
+            v = blob_task(v, v)
+
+    result = flow.run()
+    assert result.successful
+
+
+if __name__ == "__main__":
+    test_simple_flow()
